chore: move version changer command to separated module
